Split evm-test into lib and bin
